# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Al-Masriya Furniture
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProductCategory =
  (typeof ProductCategory)[keyof typeof ProductCategory];

export const ProductCategory = {
  children: "children",
  adult: "adult",
  living: "living",
  videos: "videos",
} as const;

export interface Product {
  id: number;
  name: string;
  nameEn: string;
  description: string;
  price?: number | null;
  discountPercent?: number | null;
  category: ProductCategory;
  subcategory?: string | null;
  measurements?: string | null;
  images: string[];
  videoUrl?: string | null;
  sortOrder: number;
  createdAt: string;
}

export type CreateProductBodyCategory =
  (typeof CreateProductBodyCategory)[keyof typeof CreateProductBodyCategory];

export const CreateProductBodyCategory = {
  children: "children",
  adult: "adult",
  living: "living",
  videos: "videos",
} as const;

export interface CreateProductBody {
  name: string;
  nameEn: string;
  description: string;
  price?: number | null;
  discountPercent?: number | null;
  category: CreateProductBodyCategory;
  subcategory?: string | null;
  measurements?: string | null;
  images: string[];
  videoUrl?: string | null;
  sortOrder?: number;
}

export type UpdateProductBodyCategory =
  (typeof UpdateProductBodyCategory)[keyof typeof UpdateProductBodyCategory];

export const UpdateProductBodyCategory = {
  children: "children",
  adult: "adult",
  living: "living",
  videos: "videos",
} as const;

export interface UpdateProductBody {
  name?: string;
  nameEn?: string;
  description?: string;
  price?: number | null;
  discountPercent?: number | null;
  category?: UpdateProductBodyCategory;
  subcategory?: string | null;
  measurements?: string | null;
  images?: string[];
  videoUrl?: string | null;
  sortOrder?: number;
}

export interface CategorySummary {
  category: string;
  count: number;
  label: string;
}

export type ListProductsParams = {
  category?: ListProductsCategory;
  subcategory?: string;
};

export type ListProductsCategory =
  (typeof ListProductsCategory)[keyof typeof ListProductsCategory];

export const ListProductsCategory = {
  children: "children",
  adult: "adult",
  living: "living",
  videos: "videos",
} as const;

export type DeleteProduct200 = {
  success: boolean;
};