# lib/api-zod/src/generated/api/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Al-Masriya Furniture
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all products
 */
export const ListProductsQueryParams = zod.object({
  category: zod.enum(["children", "adult", "living", "videos"]).optional(),
  subcategory: zod.coerce.string().optional(),
});

export const ListProductsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  nameEn: zod.string(),
  description: zod.string(),
  price: zod.number().nullish(),
  discountPercent: zod.number().nullish(),
  category: zod.enum(["children", "adult", "living", "videos"]),
  subcategory: zod.string().nullish(),
  measurements: zod.string().nullish(),
  images: zod.array(zod.string()),
  videoUrl: zod.string().nullish(),
  sortOrder: zod.number(),
  createdAt: zod.string(),
});
export const ListProductsResponse = zod.array(ListProductsResponseItem);

/**
 * @summary Create a product
 */
export const CreateProductBody = zod.object({
  name: zod.string(),
  nameEn: zod.string(),
  description: zod.string(),
  price: zod.number().nullish(),
  discountPercent: zod.number().nullish(),
  category: zod.enum(["children", "adult", "living", "videos"]),
  subcategory: zod.string().nullish(),
  measurements: zod.string().nullish(),
  images: zod.array(zod.string()),
  videoUrl: zod.string().nullish(),
  sortOrder: zod.number().optional(),
});

/**
 * @summary Get a single product
 */
export const GetProductParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProductResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  nameEn: zod.string(),
  description: zod.string(),
  price: zod.number().nullish(),
  discountPercent: zod.number().nullish(),
  category: zod.enum(["children", "adult", "living", "videos"]),
  subcategory: zod.string().nullish(),
  measurements: zod.string().nullish(),
  images: zod.array(zod.string()),
  videoUrl: zod.string().nullish(),
  sortOrder: zod.number(),
  createdAt: zod.string(),
});

/**
 * @summary Update a product
 */
export const UpdateProductParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateProductBody = zod.object({
  name: zod.string().optional(),
  nameEn: zod.string().optional(),
  description: zod.string().optional(),
  price: zod.number().nullish(),
  discountPercent: zod.number().nullish(),
  category: zod.enum(["children", "adult", "living", "videos"]).optional(),
  subcategory: zod.string().nullish(),
  measurements: zod.string().nullish(),
  images: zod.array(zod.string()).optional(),
  videoUrl: zod.string().nullish(),
  sortOrder: zod.number().optional(),
});

export const UpdateProductResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  nameEn: zod.string(),
  description: zod.string(),
  price: zod.number().nullish(),
  discountPercent: zod.number().nullish(),
  category: zod.enum(["children", "adult", "living", "videos"]),
  subcategory: zod.string().nullish(),
  measurements: zod.string().nullish(),
  images: zod.array(zod.string()),
  videoUrl: zod.string().nullish(),
  sortOrder: zod.number(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a product
 */
export const DeleteProductParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteProductResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get categories with product counts
 */
export const GetCategoriesSummaryResponseItem = zod.object({
  category: zod.string(),
  count: zod.number(),
  label: zod.string(),
});
export const GetCategoriesSummaryResponse = zod.array(
  GetCategoriesSummaryResponseItem,
);